finishing permissions section

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
 const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
@@ -42,9 +43,15 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id }
     // find the item
-    const item = ctx.db.query.item({where}, `{id, title}`)
+    const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
     // TODO check if user own the item
-
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission => {
+      return ['ADMIN', 'ITEMDELETE'].includes(permission);
+    })
+    if (!ownsItem && !hasPermissions) {
+      throw new Error('You don\'t have permission to delete this item');
+    }
     // delete it
     return ctx.db.mutation.deleteItem({ where }, info)
   },
@@ -154,6 +161,29 @@ const Mutations = {
     })
     // return new user
     return updatedUser;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    // check if logged in
+    if (!ctx.request.userId) {
+      throw new Error('Must be logged in.')
+    }
+    // query the current user
+    const currentUser = await ctx.db.query.user({
+      where: { id: ctx.request.userId }
+    }, info)
+    // see if they have permission to update permissions
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE'])
+    // update permissions
+    return ctx.db.mutation.updateUser({
+      data: {
+        permissions: {
+          set: args.permissions
+        }
+      },
+      where: {
+        id: args.userId
+      }
+    }, info)
   }
 };
